requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
docutils==0.20.1
idna==3.4
requests==2.31.0
urllib3==2.0.5
openai
python-decouple